Server side paginated store
Made quite a lot of changes in this one, where to start...

The store is paginated on the backend, but still sending
the whole product dataset on page load.

returned the total items count as a single number to avoid
using array.length on frontend.
Added styles for admin's categories, orders and products,
modified notifications styles.

created soe ui comps, inputs, text area and buttons.