Profile and order Pages
I added the style for the profile page as well as the
order details page. I fecthed the order details in the
backend, but the orders on the profile are still coming
from the user's list.

In the backend, I changed the enum of shipping details
and order details to adjust the shipping status and the
payment status.

I also add dynamic title thoughout the entire app, though
the product page still has difficulties to adjust the name.