Persistance
Finally, I fixed the persistence issue. now users can
login and the session will be there for 1h fot the login
token and one week for the refesh token.

The protect middleware is working too, will ave to make
a new one for admins to protect delete routes, and i
still need to see which routes i can protect with the
current one. Next optimistic updates.